20c - Added show & hide btn to Login form page password, update CRUD functions of check (incl. create & delete) & credit card voucher page.

+++ src/main/java/com/jamersc/springboot/financialhub/service/CreditCardService.java
@@ -9,10 +9,10 @@ public interface CreditCardService {
 
     List<CreditCard> findAllCreditCardRecord();
 
-    CreditCardDto findCreditCardRecordById(Integer id);
+    CreditCardDto findCreditCardRecordById(Long id);
 
     void saveCreditCardRecord(CreditCardDto creditCardDto);
 
-    void deleteCreditCardRecordById(Integer id);
+    void deleteCreditCardRecordById(Long id);
 
 }
